requirements.txt
# Core
python-dotenv==1.0.0

# LangChain & LangGraph
langgraph==0.2.45
langchain==0.3.7
langchain-openai==0.2.9
langchain-community==0.3.5
langsmith==0.1.147

# FastAPI
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Streamlit
streamlit==1.40.1

# Database
sqlalchemy==2.0.35
alembic==1.14.0

# Vector Store
faiss-cpu==1.9.0
numpy==1.26.4

# PDF Processing
pypdf==5.1.0
pdfplumber==0.11.4

# Utilities
pydantic==2.10.2
pydantic-settings==2.6.1
httpx==0.27.2
tenacity==9.0.0
